Moved token management components, CoinList to its own folder, and TransferList to its own folder
Closes #1153
