+++ apps/extension/src/ui/views/CoinList/index.tsx
